+++ packages/shared/src/hooks/usePostFeedback.ts
@@ -3,8 +3,7 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 import type { Post } from '../graphql/posts';
 import { dismissPostFeedback, UserVote } from '../graphql/posts';
 import { optimisticPostUpdateInFeed } from '../lib/feed';
-import { updatePostCache } from './usePostById';
-import { updateCachedPagePost } from '../lib/query';
+import { updateCachedPagePost, updatePostCache } from '../lib/query';
 import { ActiveFeedContext } from '../contexts/ActiveFeedContext';
 import { SharedFeedPage } from '../components/utilities';
 import type { EmptyResponse } from '../graphql/emptyResponse';
